AI/ragas_eval/requirements.txt
python-dotenv>=1.0.1
openai>=1.40.0
langchain-openai>=0.1.7
ragas>=0.1.21
datasets>=2.20.0
tqdm>=4.66.0
pandas>=2.2.2
vllm>=0.5.4; platform_system != "Windows"
